requirements.txt
numpy
Pillow
torch==2.0.0
torchvision==0.15.1
tqdm